bug fix and data type update

+++ DB2Project-be/src/main/java/it/polimi/db2/teleco_app/services/models/packagedetails/ServiceDetails.java
@@ -16,5 +16,5 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder(toBuilder = true)
 @Accessors(chain = true)
 public abstract class ServiceDetails {
-    protected Integer costMonth;
+    protected Double costMonth;
 }
